+++ changelog.md
@@ -1,4 +1,17 @@
 # Changelog
 
-**./src/output-sql-files.js**
-* Called 'saveSql' function for all tables.
+**./src/processing/object-definition/sub-location.js**
+* Rearranged property order.
+	* This corresponds to the 'StateRegion' and 'City' tables.
+	* The parent ID and item name are swapped.
+	* Index numbers of 'currentLocationName' and 'currentParentString' are swapped.
+
+---
+
+**./src/output-csv-files.js coordinateCsvWrite**
+* The 'specObject' variable is renamed to 'specObj'
+
+---
+
+**./src/output-sql-files.js coordinateSqlWrite**
+* The 'specObject' variable is renamed to 'specObj'